refactor(e2e): split advanced tests into focused modules
Break down advanced.go (664 lines) into three smaller, focused test files:
  - structured.go (222 lines): JSON schema and structured output tests
  - tools.go (275 lines): Tool/function calling tests
  - transforms.go (188 lines): Message transform tests

  Improves code organization and maintainability by keeping each file
  under 300 lines with a single, clear responsibility.
